vllm_loop: do the linear blends in the model dtype again
The damped-Euler steps and the beta blend were upcast to fp32 when _blend was
added. That is a mathematical no-op, but it changes the numerics of every
generation: a seed-matched AIME26 rerun of the fp32 version shares 0/480
generations with the bf16 run it repeats, against 359/480 for a rerun of the
same code. The runs behind the README numbers -- and the HF implementation in
loop_qwen3.py -- blend in bf16; this restores that arithmetic.

+++ vllm_loop/looped_qwen3_moe.py
@@ -93,7 +93,10 @@ def _blend(anchor, other, w, mode):
     novelty : a + w * (|a|/|b|) * b / cos(theta)         -- novelty-scaled source
     """
     if mode == "linear" or w == 0.0:
-        return ((1.0 - w) * anchor.float() + w * other.float()).to(anchor.dtype)
+        # In the operands' dtype (bf16), as in the runs behind the README numbers. An fp32
+        # upcast here is a mathematical no-op, yet it changed every one of the 480
+        # generations of a seed-matched rerun -- keep it out of the default path.
+        return ((1.0 - w) * anchor + w * other).to(anchor.dtype)
     a, b = anchor.float(), other.float()
     na = a.norm(dim=-1, keepdim=True).clamp_min(1e-6)
     nb = b.norm(dim=-1, keepdim=True).clamp_min(1e-6)
